IA/Web/requirements.txt
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
NumPy>=1.11.0
SciPy>=0.17.0
joblib>=0.11
scikit-learn>=0.21.3
pandas>=0.20.3
django-webpack-loader==0.6.0